Sharesheet - Disable delegate
The original plan was to move sharesheet into systemui. The first
step was to route through a delegate living in systemui, which 
would just launch the activity. However, the full migration has not
occurred. And more notably, there are intermittent race conditions
caused by token expiration. This comes from the framework activity token
being removed just before the target intent is launched, due to the
requirement to call finish() on resolver/chooser when complete.

Fixes: 157272342
Test: atest ChooserActivityTest ResolverActivityTest
Change-Id: I75ca19589ba78abc1d1a05a917457c151420c949