Scale down PIP when it's dragged into the dismiss target.
To do this reliably, this CL changes the motion helper logic such that all dragging and animating changes the 'temporary bounds', which are applied to PIP via userResize. When a drag/animation ends, we schedule a finishResizePip to 'actually' move PIP to its final location.

This also fixes an issue where the resize gesture could steal events during a drag.

Bug: 158247953
Fixes: 157264677
Test: manual
Change-Id: I0129a4c77d06fb9cd186eacd4a35975f54432e9d