Force app to redraw when using bounds change transaction
It can be difficult for the shell to determine whether a given
WindowContainerTransaction will actually trigger the app to redraw. If
the app doesn't redraw their bounds change transaction will be
silently dropped. We follow the same pattern used in
applySyncTransaction, and force app windows to redraw when a Task
receives a bounds change transaction.

Bug: 158600827
Test: Existing tests pass.
Change-Id: Id537fd865e9d5c7230acd7ae576990dc74d1ecfd