Fix cleanup when an app with SUSPEND_APPS is deleted
If an app with SUSPEND_APPS happened to be a system app, downgrade would
happen simultaneously on all users it was installed on. The system
should clean up suspensions on all such users in this case.
Also cleaning up the "distracting restrictions" that were being left
around earlier.

Added a new shell command to uninstall-system-updates for a specific
package.

Test: Manual.
1. Install a system app <app1> with SUSPEND_APPS and suspend another
app, <app2> with it in user 0.
2. Create another user and install an update to app1.
3. Run `adb shell pm uninstall --user all <app1>`
Verify that app2 got unsuspended in user 0.
4. Repeat 1 and then run `adb shell pm uninstall-system-updates <app1>`
Verify that app2 got unsuspended in user 0.

Bug: 139950940
Change-Id: I007b248ed522c6658d3cb48d286768c3cb3db0a1