Fix print-logs
The VERY_VERBOSE_PRINTING flag guards print-logs functionality but is
not present in the prebuilt version of the statsd module. To fix
print-logs, we remove the flag (thereby allowing print-logs to work on
user, userdebug, and eng builds), but require that the caller have root
privileges.

Test: adb shell cmd stats print-logs && adb logcat -s statsd:I (only works
once root privileges are granted)
Bug: 159982879

Change-Id: I2201fc0538876d1d6dfd6ab7abfac43cccc1102f
Merged-In: I2201fc0538876d1d6dfd6ab7abfac43cccc1102f