Update WM display info when nonOverrideDisplayInfo changes
Fixes an issue where a changed nonOverrideDisplayInfo would not be
propagated to window manager.

This causes an issue where if we change cutout shapes, the non override
display info changes, but window manager is never informed about this.

It happens to work most times, because other things trigger a refresh
of the non override display info (e.g. freezing the screen or going
to an app that overrides orientation).

Bug: 113090662
Test: Switch between cutout emulation modes a few times. Verify that the cutout change is always applied.
Change-Id: I46989e658929877775753ee453a270fa22193f11