Restore PIP to non-fullscreen bounds if it was non-fullscreen.
For the situation where we make a new stack and reparent the PIP
activity, the new stack and its task's windowConfiguration is by default
fullscreen. When PipTaskOrganizer#onTaskAppear gets called, we save the
configuration bounds, which is still fullscreen, and then restore to
that. Instead if we set the bounds of the new stack to the bounds of the
original task, then PipTaskOrganizer would get the correct configuration
when its #onTaskAppeared method is called.

Bug: 162544217
Test: Open Amazon Prime Video in freeform, play a video, enter PIP,
expand PIP - back to non-fullscreen bounds

Change-Id: I2344f618e805b2362999f476c34d8a0be62ad561