Fix checking if PiP aspect ratio has changed
PipTaskOrganizer.applyPictureInPictureParams() by design should only
return true if the PiP aspect ratio has changed. However, it's also
retrurning true if the current mPictureInPictureParams are null (not
specified). This is wrong in cases when the new PiP params do not
specify aspect ration. This lead to stange PiP behaviours (see the
test). This CL fixes it.

Bug: 165794276
Test: atest CtsSystemUiTestCases:CustomPipActionsTests#pipMenu_doesNotClose_whenUpdating_customActions
Change-Id: Ib3de6344601cd5a31b249381fd5e83275ed9c4e4