SystemConfig: Tolerate absent native lib paths.
SystemConfig.readPublicNativeLibrariesList() has some hard-coded
paths containing public.libraries*.txt files; these are not
present when this code runs outside of an Android device, for
example in a Robolectric test on host.

This CL changes the code to log a warning, rather than crash,
when one of those folders is missing or inaccessible. Note that
this increases the chance that a future breakage will be missed,
but there already was not test coverage for the method behaving
correctly.

Fixes: 169017157
Test: Checked that the following fails before this but passes
      after this CL: atest BackupManagerServiceRoboTest

Change-Id: Id8430ea68901ce60dcc855f7013bafadea823297